test fix - cast ret to c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -17,6 +17,6 @@ int main(void)
 			ret = i;
 		}
 	}
-	printf("%ld\n", ret);
+	printf("%ld\n", (int)ret);
 	return (0);
 }
